refactor(position): 위치를 담는 position을 coordinate로 세분화
